test: relocate test to userServiceTest
