+++ src/test/java/com/dope/breaking/repository/UserRepositoryTest.java
@@ -1,5 +1,6 @@
 package com.dope.breaking.repository;
 
+import com.dope.breaking.domain.user.Role;
 import com.dope.breaking.domain.user.User;
 import com.dope.breaking.service.UserService;
 import org.assertj.core.api.Assertions;
@@ -21,7 +22,7 @@ class UserRepositoryTest {
 
         //Given
         User user = new User();
-        user.signUp("URL","nickname", "01012345678","mwk300@nyu.edu","Minwu", "Kim","msg","username");
+        user.signUp("URL","nickname", "01012345678","mwk300@nyu.edu","Minwu", "Kim","msg","username", Role.USER);
 
         //When
         User savedUser = userService.save(user);
@@ -37,7 +38,7 @@ class UserRepositoryTest {
 
         //Given
         User user = new User();
-        user.signUp("URL","nickname", "01012345678","mwk300@nyu.edu","Minwu", "Kim","msg","username");
+        user.signUp("URL","nickname", "01012345678","mwk300@nyu.edu","Minwu", "Kim","msg","username",Role.USER);
 
         //When
         User savedUser = userService.save(user);
@@ -53,7 +54,7 @@ class UserRepositoryTest {
 
         //Given
         User user = new User();
-        user.signUp("URL","nickname", "01012345678","mwk300@nyu.edu","Minwu", "Kim","msg","username");
+        user.signUp("URL","nickname", "01012345678","mwk300@nyu.edu","Minwu", "Kim","msg","username",Role.USER);
 
         //When
         User savedUser = userService.save(user);
